usb-device: 'buffers::size_of' is now a 'size_t' to be consistent.
Originally 'buffers::size_of' was only passed to functions that wanted a 'uint16_t' so that's what I made it.

It's used more widely now and lives alongside 'buffers::number_of' so I think it looks better as a 'size_t'.

+++ usb-device/buffers.h
@@ -9,7 +9,7 @@ namespace buffers
 {
 
 const size_t number_of = 4;
-const uint16_t size_of = 512;
+const size_t size_of = 512;
 
 uint8_t *get_empty_buffer();
 void set_buffer_empty(uint8_t *const buffer_ptr);
